requirements-dev.txt
-i https://pypi.org/simple
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
idna==2.7
pkginfo==1.4.2
pygments==2.2.0
readme-renderer==24.0
requests-toolbelt==0.8.0
requests==2.20.0
scons-tool-loader==0.1.9
scons==3.0.1
six==1.11.0
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
